+++ models/Model.go
@@ -4,7 +4,7 @@ import "time"
 
 // Model : default Modal
 type Model struct {
-	ID        *uint      `gorm:"primary_key" json:"id,string"`
+	ID        *uint64    `gorm:"primary_key" json:"id,string"`
 	CreatedAt time.Time  `json:"created_at"`
 	UpdatedAt time.Time  `json:"updated_at"`
 	DeletedAt *time.Time `sql:"index" json:"deleted_at"`
